docker/sticky-floating-ip/src/requirements.txt
python-digitalocean==1.14.0
